pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metaharmonizer"
version = "0.3.0"
description = "Biomedical metadata harmonization: ontology mapping, schema mapping, and field suggestion"
requires-python = ">=3.10"
readme = "readme.md"
license = { text = "MIT" }
authors = [
    { name = "Sehyun Oh" },
    { name = "Abhilash Dhal", email = "adhalbiophysics@gmail.com" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    # Core data and ML
    "pandas",
    "numpy",
    "torch",
    "transformers",
    "sentence-transformers",
    "scikit-learn",
    # Vector search
    "faiss-cpu==1.11.0",
    # Fuzzy matching
    "rapidfuzz",
    "thefuzz",
    # Async HTTP and retry
    "httpx",
    "tenacity",
    "aiolimiter",
    # Utilities
    "requests",
    "python-dotenv",
    "PyYAML",
    "huggingface_hub",
    "openpyxl",
    "nltk",
    "tqdm",
]

[project.optional-dependencies]
notebook = [
    "nest-asyncio",
]
dev = [
    "pytest",
    "pytest-cov",
    "nest-asyncio",
]
eval = [
    "scipy",
]
# LLM backends are optional. Stage-4 matchers (OntoMapLLM, LLMMatcher,
# bi-encoder column inference, expand_terms_with_gemini) require Gemini.
# Semantic-clustering LLM fallback in FieldSuggester uses OpenAI.
llm-gemini = [
    "google-generativeai",
]
llm-openai = [
    "openai",
]
all = [
    "metaharmonizer[notebook,eval,llm-gemini,llm-openai]",
]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["metaharmonizer*"]

[tool.setuptools.package-data]
# Bundle the method→model YAML that lives inside metaharmonizer/models/.
"metaharmonizer.models" = ["method_model.yaml"]
# Small reference files shipped for "zero-config" first use. Accessed via
# metaharmonizer._paths.resolve_data_file() / bundled_file(). Large corpora
# and FAISS indexes are NOT bundled — users either rebuild locally or set
# METAHARMONIZER_DATA_DIR / KNOWLEDGE_DB_DIR to point at existing copies.
"metaharmonizer._bundled_data" = [
    "corpus/oncotree_code_to_name.csv",
    "schema/curated_fields.csv",
]

[tool.pytest.ini_options]
